Generate artifact files with 0644 perms (#85)
We previously created with 0777 which, depending on the caller's local
umask, would result in files created with 0755 or 0775 (assuming a
standard umask setting of 0022 or 0002).

This commit chooses 0644 as a suitably permissive mode for files. 0664
would also have been permissive, but again this is more likely to be
affected by individual umasks so is less predictable.