Create ConfigMap Artifact for chart default values
If the profile author has set default values on a Chart Artifact,
we create a ConfigMap and append it to the Helm Objects array.
The reference to the ConfigMap is then placed in the `ValuesFrom` array
in first position on the HelmRelease Artifact. Any user configuration
is then appended from index 1 onwards. This will ensure any user values
override the author's default ones.